Adding integration tests
