+++ .gitmodules
@@ -1,0 +1,3 @@
+[submodule "styra-run-sdk-tests"]
+	path = styra-run-sdk-tests
+	url = https://github.com/StyraInc/styra-run-sdk-tests.git
