+++ package.json
@@ -8,7 +8,7 @@
   "main": "./src/run-sdk.js",
   "scripts": {
     "test": "jasmine --config=tests/support/jasmine.json",
-    "integration-test-server": "node ./integration-tests/test-server"
+    "integration-test-server": "node ./tests/start-test-server"
   },
   "homepage": "https://github.com/StyraInc/styra-run-sdk-node#readme",
   "repository": {
