+++ tests/start-test-server.js
@@ -1,0 +1,3 @@
+import {startServer} from "./test-server.js";
+
+startServer(true)
